requirements.txt
Django==1.7
argparse==1.2.1
django-tastypie==0.12.0
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.8.0
wsgiref==0.1.2
